+++ packages/code-studio/src/settings/ShortcutItem.scss
