+++ packages/code-studio/src/settings/Shortcuts.ts
@@ -7,12 +7,14 @@ const NOTEBOOK = {
     name: 'Run',
     shortcut: [MODIFIER.ALT, KEY.R],
     macShortcut: [MODIFIER.OPTION, KEY.R],
+    isEditable: false,
   }),
   RUN_SELECTED: ShortcutRegistry.createAndAdd({
     id: 'NOTEBOOK.RUN_SELECTED',
     name: 'Run Selected',
     shortcut: [MODIFIER.ALT, MODIFIER.SHIFT, KEY.R],
     macShortcut: [MODIFIER.OPTION, MODIFIER.SHIFT, KEY.R],
+    isEditable: false,
   }),
   FIND: ShortcutRegistry.createAndAdd({
     id: 'NOTEBOOK.FIND',
@@ -39,12 +41,9 @@ const FILE_EXPLORER = {
   }),
 };
 
-const CHART = { name: 'Chart' };
-
 const SHORTCUTS = {
   NOTEBOOK,
   FILE_EXPLORER,
-  CHART,
 };
 
 export default SHORTCUTS;
