+++ packages/code-studio/src/styleguide/ContextMenus.jsx
@@ -34,7 +34,7 @@ class ContextMenus extends Component {
           console.log(`Logging a message${suffix}!`);
         },
         shortcut: new Shortcut({
-          id: 'STYLEGUIDE.LOW_MESSAGE',
+          id: 'STYLEGUIDE.LOG_MESSAGE',
           name: 'Log Message',
           shortcut: [MODIFIER.CTRL, KEY.L],
           macShortcut: [MODIFIER.CMD, KEY.L],
